requirements.txt
google-genai
feedparser
requests
psycopg2-binary
youtube-transcript-api==0.6.3